Split sail_wrapper_128r into two files
In preparation for the 32-bit format
